# [TorchClassifierAgent]Don't share optimizer if we don't have it