@dexterju 2020-04-21
```

tests/test_transformers.py .                                                                                                                            [100%]

================================================================== slowest 10 test durations ==================================================================
19.82s call     tests/test_transformers.py::TestClassifier::test_eval_classifier
```